🚚  rename :: 공용 컴포넌트 atoms 폴더로 분리
