+++ src/refactoring/components/admin/CouponList.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { ItemListTemplate } from '../templates';
 import { Coupon } from '../../../types';
-import { CouponItem } from './CouponItem';
+import { CouponItem } from '../atoms'
 import { TitleLabel } from "../atoms";
 
 type CouponListProps = {
