+++ src/refactoring/components/admin/index.ts
@@ -1,8 +1,6 @@
 export {CouponForm} from './CouponForm'
-export {CouponItem} from './CouponItem'
 export {CouponList} from './CouponList'
 export {ProductForm} from './ProductForm'
 export {ProductItem} from './ProductItem'
 export {ProductDetailInfo} from './ProductDetailInfo'
 export {ProductDetail} from './ProductDetail'
-export {DiscountList} from './DiscountList'
