+++ src/refactoring/components/atoms/index.ts
@@ -3,4 +3,6 @@ export { Select } from './Select';
 export { Label } from './Label';
 export { LabelButton } from './LabelButton';
 export { Input } from './Input';
+export { CouponItem } from './CouponItem'
 export { TitleLabel } from './TitleLabel'
+export {DiscountItem} from './DiscountItem'
