+++ src/refactoring/pages/index.ts
@@ -1,1 +1,3 @@
-export * from './MainPage'
+export {MainPage} from './MainPage'
+export {AdminPage} from './AdminPage'
+export {CartPage} from './CartPage'
